Create new raster paint layer module and factor out BaseTimeseriesPropos
