+++ app/scripts/components/common/map/style-generators/cmr-timeseries.tsx
@@ -1,27 +1,9 @@
 import React from 'react';
-import { BaseGeneratorParams } from '../types';
-import { ZarrPaintLayer } from './zarr-timeseries';
+import { CMRTimeseriesProps } from '../types';
+import { RasterPaintLayer } from './raster-paint-layer';
 import { useCMR } from './hooks';
-import { ActionStatus } from '$utils/status';
 
-interface AssetUrlReplacement {
-  from: string;
-  to: string;
-}
-
-export interface CMRTimeseriesProps extends BaseGeneratorParams {
-  id: string;
-  stacCol: string;
-  date?: Date;
-  sourceParams?: Record<string, any>;
-  stacApiEndpoint?: string;
-  tileApiEndpoint?: string;
-  assetUrlReplacements?: AssetUrlReplacement;
-  zoomExtent?: number[];
-  onStatusChange?: (result: { status: ActionStatus; id: string }) => void;
-}
-
-export function CMRTimeseries(props:CMRTimeseriesProps) {
+export function CMRTimeseries(props: CMRTimeseriesProps) {
   const {
     id,
     stacCol,
@@ -33,5 +15,5 @@ export function CMRTimeseries(props:CMRTimeseriesProps) {
 
   const stacApiEndpointToUse = stacApiEndpoint?? process.env.API_STAC_ENDPOINT;
   const assetUrl = useCMR({ id, stacCol, stacApiEndpointToUse, date, assetUrlReplacements, stacApiEndpoint, onStatusChange });
-  return <ZarrPaintLayer {...props} assetUrl={assetUrl} />;
+  return <RasterPaintLayer {...props} assetUrl={assetUrl} />;
 }
